Initial commit
This commit introduces a new board qemu-x86_64-debug.

This board is just a copy of the qemu-x86_64 board with debug symbols
and facilities for the Linux kernel.
It also remove the KASLR from the cmdline and wait for a gdb connection
on :1234 port.

A minimal example kernel module is also provided.

To avoid a compilation error with binutils >= 2.35.2 and qemu 5.2.0, a
qemu patch has been added.